requirements.txt
accelerate==0.19.0
diffusers==0.16.1
einops==0.6.1
#flash_attn==1.0.5
Flask==2.3.2
ftfy==6.1.1
labml==0.4.162
labml_nn==0.4.133
ldm==0.1.3
lbry-libtorrent
matplotlib==3.7.1
numpy==1.24.3
omegaconf==2.3.0
Pillow==9.0.0
pytorch_lightning
Requests==2.30.0
scipy==1.10.1
torch==2.0.1
tqdm==4.65.0
transformers==4.29.2
fire
jupyter-runner
torchtext==0.6.0
libtorrent
